AI_Project/requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
numpy 
pandas
matplotlib
scikit-learn
seaborn
langgraph
python-dotenv
pypdf
langchain-text-splitters
chromadb
faiss-cpu
streamlit
google-generativeai
langchain_google_genai
PyPDF2
google-search-results
requests
bs4